Shifted grid2geo, geo2grid, llh2xyz, and xyz2llh from transform.py to convert.py
